move all sidebar classes to new sidebar app/dir
