+++ e2e/app/page/notebook-page.ts
@@ -3,7 +3,7 @@ import { ElementHandle, Frame, Page } from 'puppeteer';
 import { getPropValue } from 'utils/element-utils';
 import { waitForDocumentTitle, waitForNumericalString, waitWhileLoading } from 'utils/waits-utils';
 import { ResourceCard } from 'app/text-labels';
-import RuntimePanel, { StartStopIconState } from 'app/component/runtime-panel';
+import RuntimePanel, { StartStopIconState } from 'app/sidebar/runtime-panel';
 import NotebookCell, { CellType } from './notebook-cell';
 import NotebookDownloadModal from 'app/modal/notebook-download-modal';
 import NotebookPreviewPage from './notebook-preview-page';
