+++ e2e/tests/nightly/genomic-extraction-to-vcf.spec.ts
@@ -13,9 +13,9 @@ import CohortActionsPage from 'app/page/cohort-actions-page';
 import { makeRandomName, makeWorkspaceName } from 'utils/str-utils';
 import GenomicsVariantExtractConfirmationModal from 'app/modal/genomic-extract-confirmation-modal';
 import ExportToNotebookModal from 'app/modal/export-to-notebook-modal';
-import RuntimePanel, { AutoPauseIdleTime, ComputeType } from 'app/component/runtime-panel';
+import RuntimePanel, { AutoPauseIdleTime, ComputeType } from 'app/sidebar/runtime-panel';
 import { logger } from 'libs/logger';
-import GenomicExtractionsSidebar from 'app/component/genomic-extractions-sidebar';
+import GenomicExtractionsSidebar from 'app/sidebar/genomic-extractions-sidebar';
 import { Page } from 'puppeteer';
 import { takeScreenshot } from 'utils/save-file-utils';
 import expect from 'expect';
